requirements.txt
azure-iot-device
CounterFit
counterfit-connection
counterfit-shims-grove
flake8
mockito
pytest
radon
py7zr